+++ CHANGELOG.md
@@ -4,9 +4,7 @@ All notable changes to Orynivo are documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
-## [Unreleased]
-
-### Added
+## [0.36.5] - 2026-08-12
 
 ### Fixed